pyproject.toml
[tool.poetry]
name = "moviereservation"
version = "0.1.0"
description = ""
authors = ["Aron Sadegh  <aronesadegh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.1.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
